Lab2/requirements.txt
yfinance==0.2.54
pandas>=2.0
requests
certifi
apache-airflow-providers-snowflake
